requirements.txt
pytest==8.4.2
coverage==6.5.0
pytest-mock
requests
flask
behave
pymongo-amplidata